\name{letter421-package}
\alias{letter421-package}
\alias{letter421}
\docType{package}
\title{
\packageTitle{letter421}
}
\description{
\packageDescription{letter421}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{letter421}
\packageIndices{letter421}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{letter421}

Maintainer: \packageMaintainer{letter421}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
